+++ src/main/java/com/odontovision/Java_API/services/ProcedimentoService.java
@@ -5,6 +5,7 @@ import com.odontovision.Java_API.repositories.ProcedimentoRepository;
 import com.odontovision.Java_API.dtos.ProcedimentoDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -14,12 +15,11 @@ public class ProcedimentoService {
     @Autowired
     private ProcedimentoRepository procedimentoRepository;
 
-    public Procedimento cadastrarProcedimento(ProcedimentoDTO dto) {
+    public Procedimento salvarProcedimento(ProcedimentoDTO dto) {
         Procedimento procedimento = new Procedimento();
         procedimento.setNomeProcedimento(dto.getNomeProcedimento());
         procedimento.setDescricao(dto.getDescricao());
         procedimento.setCusto(dto.getCusto());
-
         return procedimentoRepository.save(procedimento);
     }
 
@@ -27,11 +27,14 @@ public class ProcedimentoService {
         return procedimentoRepository.findAll();
     }
 
-    public Optional<Procedimento> buscarPorId(Long id) {
+    public Optional<Procedimento> buscarProcedimentoPorId(Long id) {
         return procedimentoRepository.findById(id);
     }
 
     public void excluirProcedimento(Long id) {
+        if (!procedimentoRepository.existsById(id)) {
+            throw new RuntimeException("Procedimento não encontrado");
+        }
         procedimentoRepository.deleteById(id);
     }
 }
